+++ brostar_api/settings.py
@@ -299,6 +299,8 @@ CELERY_IMPORTS = ("api.tasks",)
 
 # TODO: fix celery env settings
 CELERY_BROKER_URL = "redis://redis:6379/0"
+# Use Redis as the result backend (required for chords)
+CELERY_RESULT_BACKEND = "redis://redis:6379/0"
 # Auto-expire results after 1 day
 CELERY_RESULT_EXPIRES = 60 * 60 * 24
 
@@ -307,6 +309,7 @@ CELERY_TASK_QUEUES = (
     # Define all your requested queues
     Queue("default", Exchange("default"), routing_key="default"),
     Queue("upload", Exchange("upload"), routing_key="upload"),
+    Queue("gld_import", Exchange("gld_import"), routing_key="gld_import"),
 )
 
 # Default queue if not specified